fix: Address PR comments.
1. Pagination fix: fetch more data until a have a page with fewer
   elements than the page size.

2. Applied the suggested diff for len(CustomFields)

3.  DeleteL2PVN now takes int64 and does the bounds check internally.

4. L2VPNClaimSpec.Identifier removed. To specify a single VNI set start
   and end to the same value.

5. getLeaseLockerNSNandOwner now guards against an empty
   `OwnerReferences` slice internally

+++ pkg/netbox/api/l2vpn_test.go
@@ -222,7 +222,7 @@ func TestL2VPN(t *testing.T) {
 
 		compositeClient := &NetboxCompositeClient{clientV4: &NetboxClientV4{VpnAPI: mockVpnAPI}}
 
-		err := compositeClient.DeleteL2VPN(context.TODO(), L2VPNId)
+		err := compositeClient.DeleteL2VPN(context.TODO(), int64(L2VPNId))
 		AssertNil(t, err)
 	})
 
@@ -235,7 +235,7 @@ func TestL2VPN(t *testing.T) {
 
 		compositeClient := &NetboxCompositeClient{clientV4: &NetboxClientV4{VpnAPI: mockVpnAPI}}
 
-		err := compositeClient.DeleteL2VPN(context.TODO(), L2VPNId)
+		err := compositeClient.DeleteL2VPN(context.TODO(), int64(L2VPNId))
 		AssertNil(t, err)
 	})
 
@@ -248,7 +248,7 @@ func TestL2VPN(t *testing.T) {
 
 		compositeClient := &NetboxCompositeClient{clientV4: &NetboxClientV4{VpnAPI: mockVpnAPI}}
 
-		err := compositeClient.DeleteL2VPN(context.TODO(), L2VPNId)
+		err := compositeClient.DeleteL2VPN(context.TODO(), int64(L2VPNId))
 		AssertError(t, err, "failed to delete l2vpn from netbox: status 400, body: ")
 	})
 }
